fix: handle case where an epoch is not finalized
If a snapshotter submits snapshots for an epoch which doesn't get finalized, self-healing kicks-in and tries to fill an empty payload which causes 2 entries in redis at same score.
In order to avoid this special case, the existing entry is being removed when status from consensus is returned as not finalized.