fix: handle scenario due to consensus service error response during self-healing
A scenario where snapshot is submitted, but finalization has not happened and during self-healing,
an error response is received from consensus service a null snapshot gets inserted in dag block but the redis
payloads Zset will still have a stale entry of the older snapshot submitted.
Handling this case so that the older entry is removed before a null entry is inserted.